+++ src/cogent3/util/dict_array.py
@@ -30,11 +30,11 @@ from typing import TYPE_CHECKING, Any, Protocol, Self, SupportsInt, TypeVar, cas
 
 import numpy
 import numpy.typing as npt
+from scinexus.deserialise import get_class, register_deserialiser
 from scinexus.io_util import atomic_write
 from scinexus.misc import get_object_provenance
 
 from cogent3._version import __version__
-from cogent3.util.deserialise import get_class, register_deserialiser
 from cogent3.util.io import PathType
 
 if TYPE_CHECKING:  # pragma: no cover
